Use result_type in insertConstant
